Novel Forest Dataset and Model

The dataset and model used in Lai et al. (2021), who examined the diversity--environment relationships of native and exotic species in two types of post-cultivation novel secondary forests in Singapore. The generalised linear mixed-effect model was fitted via brms::brm so this package is required to properly view the model object.

novelforest

Format

A brms model output of class brmsfit, which is a list containing the input data and other slots that store the model components. Here, we will only briefly describe

how the data can be extracted and inspected

the model summary

Please refer to ?brms::brmsfit or ?brms::brm for more information on the model object.

To access the dataset, use novelforest$data. The dataset contains the following response variables:

SD_N_0

first-order native taxonomic diversity, i.e., species richness

SD_N_2

second-order native taxonomic diversity, i.e., inverse Simpson index

SD_E_0

first-order exotic taxonomic diversity

SD_E_2

second-order exotic taxonomic diversity

FD_N_0

first-order native functional diversity

FD_N_2

second-order native functional diversity

FD_E_0

first-order exotic functional diversity

FD_E_2

second-order exotic functional diversity,

and the following explanatory variables (and measurement units if you backtransform them using backtransform):

dist

Distance to old-growth forests (m)

size

Patch area (km^2)

nitrogen

Total soil nitrogen (mg/kg)

phosphorous

Total extractable soil phosphorous (mg/kg)

potassium

Total extractable soil potassium (mg/kg)

patch

Forest patch ID

Note that all explanatory variables were log-transformed and standardised to zero mean and unit standard deviations. See Lai et al. (2021) for more details on model building and data collection.

References

Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W. and Chong, K.Y. (2021) Decoupled responses of native and exotic tree diversities to distance from old-growth forest and soil phosphorous in novel secondary forests. Applied Vegetation Science, 24, e12548. https://doi.org/10.1111/avsc.12548

See also

Examples

# To access the dataset
dat <- novelforest$data
summary(dat)

#>      SD_N_0            dist              size            nitrogen      
#>  Min.   : 0.000   Min.   :-3.0453   Min.   :-1.7103   Min.   :-3.0219  
#>  1st Qu.: 2.000   1st Qu.:-0.4258   1st Qu.:-1.0215   1st Qu.:-0.5493  
#>  Median : 4.000   Median : 0.1160   Median : 0.0436   Median : 0.0132  
#>  Mean   : 4.402   Mean   : 0.0000   Mean   : 0.0000   Mean   : 0.0000  
#>  3rd Qu.: 6.000   3rd Qu.: 0.4551   3rd Qu.: 0.8288   3rd Qu.: 0.6091  
#>  Max.   :12.000   Max.   : 1.4943   Max.   : 1.7053   Max.   : 3.0982  
#>                                                                        
#>   phosphorous        potassium            patch        SD_N_2     
#>  Min.   :-1.9614   Min.   :-2.44769   AD     : 5   Min.   :0.000  
#>  1st Qu.:-0.6397   1st Qu.:-0.56177   AW     : 5   1st Qu.:1.839  
#>  Median :-0.1474   Median :-0.01242   BBE    : 5   Median :2.814  
#>  Mean   : 0.0000   Mean   : 0.00000   BBH    : 5   Mean   :2.980  
#>  3rd Qu.: 0.3470   3rd Qu.: 0.70705   BBN    : 5   3rd Qu.:4.000  
#>  Max.   : 3.1607   Max.   : 2.00352   BBT    : 5   Max.   :8.167  
#>                                       (Other):67                  
#>      SD_E_0          SD_E_2          FD_N_0            FD_N_2       
#>  Min.   :0.000   Min.   :0.000   Min.   :  0.000   Min.   :  0.000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:  6.344   1st Qu.:  6.344  
#>  Median :2.000   Median :1.220   Median : 38.604   Median : 28.286  
#>  Mean   :2.031   Mean   :1.539   Mean   : 70.365   Mean   : 39.320  
#>  3rd Qu.:3.000   3rd Qu.:1.800   3rd Qu.: 95.374   3rd Qu.: 56.151  
#>  Max.   :6.000   Max.   :5.333   Max.   :420.347   Max.   :204.312  
#>                                                                     
#>      FD_E_0           FD_E_2      
#>  Min.   : 0.000   Min.   : 0.000  
#>  1st Qu.: 0.000   1st Qu.: 0.000  
#>  Median : 6.328   Median : 6.328  
#>  Mean   :11.508   Mean   : 8.621  
#>  3rd Qu.:18.938   3rd Qu.:12.114  
#>  Max.   :95.706   Max.   :79.645  
#>                                   

# To summarise the model parameters
summary(novelforest)

#>  Family: MV(hurdle_gamma, hurdle_gamma, hurdle_gamma, hurdle_gamma, hurdle_gamma, hurdle_gamma, hurdle_gamma, hurdle_gamma) 
#>   Links: mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity
#>          mu = log; shape = identity; hu = identity 
#> Formula: SD_N_0 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          SD_N_2 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          SD_E_0 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          SD_E_2 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          FD_N_0 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          FD_N_2 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          FD_E_0 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>          FD_E_2 ~ dist + size + nitrogen + phosphorous + potassium + (1 | 1 | patch) 
#>    Data: novel_D_scaled_long (Number of observations: 97) 
#> Samples: 4 chains, each with iter = 2000; warmup = 1000; thin = 1;
#>          total post-warmup samples = 4000
#> 
#> Group-Level Effects: 
#> ~patch (Number of levels: 20) 
#>                                    Estimate Est.Error l-95% CI u-95% CI Rhat
#> sd(SDN0_Intercept)                     0.24      0.07     0.10     0.40 1.00
#> sd(SDN2_Intercept)                     0.17      0.07     0.04     0.32 1.00
#> sd(SDE0_Intercept)                     0.30      0.07     0.19     0.47 1.00
#> sd(SDE2_Intercept)                     0.27      0.06     0.16     0.41 1.00
#> sd(FDN0_Intercept)                     0.53      0.16     0.26     0.88 1.00
#> sd(FDN2_Intercept)                     0.43      0.12     0.22     0.70 1.00
#> sd(FDE0_Intercept)                     0.51      0.17     0.22     0.89 1.00
#> sd(FDE2_Intercept)                     0.45      0.13     0.23     0.75 1.00
#> cor(SDN0_Intercept,SDN2_Intercept)     0.42      0.26    -0.16     0.84 1.00
#> cor(SDN0_Intercept,SDE0_Intercept)    -0.13      0.25    -0.60     0.38 1.00
#> cor(SDN2_Intercept,SDE0_Intercept)    -0.01      0.27    -0.51     0.51 1.00
#> cor(SDN0_Intercept,SDE2_Intercept)    -0.12      0.25    -0.59     0.39 1.00
#> cor(SDN2_Intercept,SDE2_Intercept)     0.00      0.26    -0.51     0.51 1.00
#> cor(SDE0_Intercept,SDE2_Intercept)     0.62      0.19     0.16     0.91 1.00
#> cor(SDN0_Intercept,FDN0_Intercept)     0.51      0.24    -0.03     0.88 1.00
#> cor(SDN2_Intercept,FDN0_Intercept)     0.40      0.27    -0.19     0.82 1.00
#> cor(SDE0_Intercept,FDN0_Intercept)    -0.13      0.25    -0.61     0.38 1.00
#> cor(SDE2_Intercept,FDN0_Intercept)    -0.17      0.25    -0.62     0.33 1.00
#> cor(SDN0_Intercept,FDN2_Intercept)     0.50      0.23    -0.03     0.87 1.00
#> cor(SDN2_Intercept,FDN2_Intercept)     0.46      0.25    -0.13     0.85 1.00
#> cor(SDE0_Intercept,FDN2_Intercept)    -0.06      0.25    -0.52     0.42 1.00
#> cor(SDE2_Intercept,FDN2_Intercept)    -0.09      0.25    -0.56     0.39 1.00
#> cor(FDN0_Intercept,FDN2_Intercept)     0.56      0.22     0.05     0.89 1.00
#> cor(SDN0_Intercept,FDE0_Intercept)    -0.11      0.27    -0.61     0.41 1.00
#> cor(SDN2_Intercept,FDE0_Intercept)     0.02      0.28    -0.52     0.57 1.00
#> cor(SDE0_Intercept,FDE0_Intercept)     0.43      0.26    -0.12     0.85 1.00
#> cor(SDE2_Intercept,FDE0_Intercept)     0.40      0.26    -0.15     0.83 1.00
#> cor(FDN0_Intercept,FDE0_Intercept)    -0.08      0.26    -0.57     0.43 1.00
#> cor(FDN2_Intercept,FDE0_Intercept)     0.00      0.26    -0.50     0.51 1.00
#> cor(SDN0_Intercept,FDE2_Intercept)    -0.10      0.26    -0.59     0.42 1.00
#> cor(SDN2_Intercept,FDE2_Intercept)     0.06      0.28    -0.48     0.60 1.00
#> cor(SDE0_Intercept,FDE2_Intercept)     0.47      0.25    -0.09     0.86 1.00
#> cor(SDE2_Intercept,FDE2_Intercept)     0.46      0.24    -0.07     0.84 1.00
#> cor(FDN0_Intercept,FDE2_Intercept)    -0.09      0.26    -0.58     0.43 1.00
#> cor(FDN2_Intercept,FDE2_Intercept)     0.02      0.26    -0.49     0.49 1.00
#> cor(FDE0_Intercept,FDE2_Intercept)     0.50      0.23    -0.02     0.87 1.00
#>                                    Bulk_ESS Tail_ESS
#> sd(SDN0_Intercept)                     2125     1672
#> sd(SDN2_Intercept)                     1914     1528
#> sd(SDE0_Intercept)                     2713     2767
#> sd(SDE2_Intercept)                     2833     2823
#> sd(FDN0_Intercept)                     2602     2718
#> sd(FDN2_Intercept)                     2917     2971
#> sd(FDE0_Intercept)                     2623     2121
#> sd(FDE2_Intercept)                     3251     2993
#> cor(SDN0_Intercept,SDN2_Intercept)     2888     3033
#> cor(SDN0_Intercept,SDE0_Intercept)     2014     2695
#> cor(SDN2_Intercept,SDE0_Intercept)     1698     2572
#> cor(SDN0_Intercept,SDE2_Intercept)     1935     2603
#> cor(SDN2_Intercept,SDE2_Intercept)     1886     2750
#> cor(SDE0_Intercept,SDE2_Intercept)     2898     3182
#> cor(SDN0_Intercept,FDN0_Intercept)     2595     3276
#> cor(SDN2_Intercept,FDN0_Intercept)     2542     2540
#> cor(SDE0_Intercept,FDN0_Intercept)     3879     3369
#> cor(SDE2_Intercept,FDN0_Intercept)     4459     3802
#> cor(SDN0_Intercept,FDN2_Intercept)     2611     3069
#> cor(SDN2_Intercept,FDN2_Intercept)     2300     2276
#> cor(SDE0_Intercept,FDN2_Intercept)     4164     3662
#> cor(SDE2_Intercept,FDN2_Intercept)     4088     3814
#> cor(FDN0_Intercept,FDN2_Intercept)     3341     3584
#> cor(SDN0_Intercept,FDE0_Intercept)     2904     3106
#> cor(SDN2_Intercept,FDE0_Intercept)     2607     2455
#> cor(SDE0_Intercept,FDE0_Intercept)     3099     2966
#> cor(SDE2_Intercept,FDE0_Intercept)     3204     3135
#> cor(FDN0_Intercept,FDE0_Intercept)     3833     3763
#> cor(FDN2_Intercept,FDE0_Intercept)     3759     3011
#> cor(SDN0_Intercept,FDE2_Intercept)     2648     3128
#> cor(SDN2_Intercept,FDE2_Intercept)     2558     3065
#> cor(SDE0_Intercept,FDE2_Intercept)     3306     3214
#> cor(SDE2_Intercept,FDE2_Intercept)     3617     3442
#> cor(FDN0_Intercept,FDE2_Intercept)     3793     3724
#> cor(FDN2_Intercept,FDE2_Intercept)     3737     3915
#> cor(FDE0_Intercept,FDE2_Intercept)     3018     3402
#> 
#> Population-Level Effects: 
#>                  Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> SDN0_Intercept       1.45      0.08     1.30     1.59 1.00     2471     3140
#> SDN2_Intercept       1.09      0.06     0.97     1.21 1.00     3526     3150
#> SDE0_Intercept       0.70      0.08     0.53     0.87 1.00     2589     2858
#> SDE2_Intercept       0.43      0.07     0.30     0.58 1.00     2877     3001
#> FDN0_Intercept       4.09      0.16     3.77     4.40 1.00     2631     2617
#> FDN2_Intercept       3.55      0.14     3.29     3.82 1.00     2665     2917
#> FDE0_Intercept       2.79      0.19     2.42     3.15 1.00     2981     2769
#> FDE2_Intercept       2.51      0.16     2.18     2.83 1.00     3037     2925
#> SDN0_dist           -0.32      0.08    -0.47    -0.17 1.00     3516     3261
#> SDN0_size           -0.01      0.08    -0.17     0.15 1.00     3212     3140
#> SDN0_nitrogen       -0.10      0.07    -0.24     0.04 1.00     4239     3171
#> SDN0_phosphorous    -0.02      0.06    -0.13     0.11 1.00     4757     3294
#> SDN0_potassium      -0.04      0.07    -0.18     0.10 1.00     4524     2890
#> SDN2_dist           -0.29      0.07    -0.42    -0.16 1.00     4268     3472
#> SDN2_size           -0.01      0.07    -0.13     0.12 1.00     3869     2990
#> SDN2_nitrogen       -0.07      0.06    -0.19     0.05 1.00     4631     3654
#> SDN2_phosphorous    -0.01      0.05    -0.11     0.10 1.00     6167     3247
#> SDN2_potassium      -0.01      0.06    -0.13     0.11 1.00     5330     3192
#> SDE0_dist            0.11      0.09    -0.05     0.29 1.00     2932     2876
#> SDE0_size           -0.04      0.09    -0.22     0.14 1.00     2566     2859
#> SDE0_nitrogen        0.02      0.07    -0.11     0.15 1.00     4985     3380
#> SDE0_phosphorous     0.20      0.06     0.07     0.32 1.00     4841     3369
#> SDE0_potassium      -0.00      0.07    -0.14     0.14 1.00     5077     3393
#> SDE2_dist            0.05      0.08    -0.09     0.20 1.00     2915     3138
#> SDE2_size           -0.05      0.08    -0.20     0.11 1.00     2461     2813
#> SDE2_nitrogen        0.09      0.06    -0.02     0.20 1.00     5029     3279
#> SDE2_phosphorous     0.12      0.05     0.01     0.23 1.00     4945     2635
#> SDE2_potassium      -0.05      0.06    -0.17     0.07 1.00     5274     2724
#> FDN0_dist           -0.53      0.17    -0.86    -0.20 1.00     3611     3361
#> FDN0_size           -0.03      0.17    -0.36     0.31 1.00     3199     2842
#> FDN0_nitrogen       -0.18      0.14    -0.45     0.09 1.00     4723     3405
#> FDN0_phosphorous    -0.12      0.12    -0.36     0.12 1.00     5512     3337
#> FDN0_potassium      -0.10      0.14    -0.38     0.18 1.00     5207     3438
#> FDN2_dist           -0.48      0.14    -0.75    -0.21 1.00     4100     3329
#> FDN2_size           -0.01      0.14    -0.27     0.26 1.00     3256     2860
#> FDN2_nitrogen       -0.14      0.11    -0.35     0.07 1.00     4907     3221
#> FDN2_phosphorous    -0.09      0.10    -0.28     0.10 1.00     5491     2921
#> FDN2_potassium      -0.07      0.11    -0.29     0.14 1.00     4714     3254
#> FDE0_dist            0.06      0.18    -0.29     0.44 1.00     3838     3153
#> FDE0_size            0.17      0.18    -0.17     0.53 1.00     4030     3372
#> FDE0_nitrogen       -0.03      0.15    -0.32     0.26 1.00     5117     3221
#> FDE0_phosphorous     0.04      0.15    -0.26     0.33 1.00     5056     3192
#> FDE0_potassium      -0.00      0.18    -0.36     0.35 1.00     4897     3226
#> FDE2_dist            0.04      0.16    -0.27     0.36 1.00     3728     2875
#> FDE2_size            0.14      0.15    -0.17     0.42 1.00     3359     3036
#> FDE2_nitrogen       -0.00      0.12    -0.25     0.24 1.00     5668     3059
#> FDE2_phosphorous     0.03      0.12    -0.22     0.27 1.00     5029     3164
#> FDE2_potassium       0.00      0.15    -0.28     0.29 1.00     6019     2963
#> 
#> Family Specific Parameters: 
#>            Estimate Est.Error l-95% CI u-95% CI Rhat Bulk_ESS Tail_ESS
#> shape_SDN0     4.85      0.77     3.48     6.48 1.00     5826     2784
#> shape_SDN2     6.01      0.93     4.33     7.99 1.00     5631     3015
#> shape_SDE0     5.69      0.89     4.10     7.62 1.00     7002     3158
#> shape_SDE2     8.11      1.29     5.79    10.87 1.00     7373     2878
#> shape_FDN0     1.56      0.25     1.13     2.08 1.00     6142     2904
#> shape_FDN2     2.28      0.36     1.63     3.06 1.00     6048     3098
#> shape_FDE0     1.86      0.36     1.23     2.65 1.00     5540     2772
#> shape_FDE2     2.66      0.53     1.73     3.80 1.00     5413     2902
#> hu_SDN0        0.07      0.03     0.03     0.13 1.00     7877     2644
#> hu_SDN2        0.07      0.03     0.03     0.13 1.00     6826     2724
#> hu_SDE0        0.05      0.02     0.02     0.10 1.00     9744     2676
#> hu_SDE2        0.05      0.02     0.02     0.10 1.00     9363     3035
#> hu_FDN0        0.15      0.04     0.09     0.23 1.00     8968     2818
#> hu_FDN2        0.15      0.04     0.09     0.23 1.00     9428     2845
#> hu_FDE0        0.39      0.05     0.30     0.49 1.00     9173     2550
#> hu_FDE2        0.39      0.05     0.30     0.49 1.00     9010     2500
#> 
#> Samples were drawn using sampling(NUTS). For each parameter, Bulk_ESS
#> and Tail_ESS are effective sample size measures, and Rhat is the potential
#> scale reduction factor on split chains (at convergence, Rhat = 1).