Novel Forest Raw Dataset

The raw data used in Lai et al. (2021) and Neo et al. (2017).

novelforest_data$trees
novelforest_data$pres
novelforest_data$plots
novelforest_data$patches

Format

A list of four data.frame:

trees

a table of woody stems measured, with treeID as the unique identifier

pres

a table of presences of vascular plant species in each plot

plots

a sf table of the environmental measurements and landscape characteristics for each plot, with plotID as the unique identifier; coordinates are in WGS84 (EPSG:4326)

patches

a table of the patches and their characteristics, with patchID as the unique identifier

With the following variables:

patch

Forest patch identifier

plot

Plot identifier

UID

Stem identifier

species

Species name following Chong et al. (2011)

stem

Indicator column denoting whether a stem is the main trunk (=1) of an individual tree, or otherwise (=0)

dbh_2011

Diameter-at-breast-height (cm) measured in year 2011.

canopy

% Canopy cover, averaged across five readings per plot

litter

Leaf litter depth / cm, averaged across five readings per plot

nitrogen

Soil total Nitrogen /g^-1 kg^-1

phosphorous

Soil extractable Phosphorous /mg^-1 kg^-1

potassium

Soil extractable Potassium /mg^-1 kg^-1

dist

Distance to nearest old-growth forest /km

geometry

Plot location as an sfc_POINT list-column in WGS84 (EPSG:4326), created as the centroid of plot corner marker locations

type

Type of forest; ABA: abandoned-land forest type, WAS: waste-woodland forest type

size

Size (i.e., area) of forest patch /ha

See Lai et al. (2021) or Neo et al. (2017) for more detail on data collection.

References

Chong, K. Y., Tan, H. T. W. and Corlett, R. T. (2011). A summary of the total vascular plant flora of Singapore. Gardens' Bulletin Singapore, 63, 197-204.

Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W. and Chong, K.Y. (2021) Decoupled responses of native and exotic tree diversities to distance from old-growth forest and soil phosphorous in novel secondary forests. Applied Vegetation Science, 24, e12548. doi:10.1111/avsc.12548

Neo, L., Yee, A. T. K., Chong, K. Y., Kee, C. Y. and Tan, H. T. W. (2017). Vascular plant species richness and composition in two types of post-cultivation tropical secondary forest. Applied Vegetation Science, 20, 692-701. doi:10.1111/avsc.12322

Examples

# Tree community matrix (by stem counts)
comm_stem <- xtabs(~ plot + species, data = novelforest_data$trees)
comm_stem[1:10, 1:4]
#>      species
#> plot  Acacia auriculiformis Adenanthera pavonina Adinandra dumosa
#>   AD1                     0                    0                0
#>   AD2                     0                    0                0
#>   AD3                     0                    0                0
#>   AD4                     0                    0                0
#>   AD5                     0                    0                0
#>   AW1                     0                    0                6
#>   AW2                     9                    0                0
#>   AW3                     0                    0                9
#>   AW4                     4                    0                0
#>   AW5                    44                    0                0
#>      species
#> plot  Aidia densiflora
#>   AD1                0
#>   AD2                0
#>   AD3                0
#>   AD4                0
#>   AD5                0
#>   AW1                0
#>   AW2                0
#>   AW3                0
#>   AW4                0
#>   AW5                0

# Species occurrence matrix (of all vascular plants)
comm_all <- xtabs(~ plot + species, data = novelforest_data$pres)
comm_all[1:10, 1:4]
#>      species
#> plot  Abroma augusta  Abrus precatorius  Acacia auriculiformis Acacia concinna
#>   AD1               0                  0                     0               0
#>   AD2               0                  0                     0               0
#>   AD3               0                  0                     0               0
#>   AD4               0                  0                     0               0
#>   AD5               0                  0                     0               0
#>   AW1               0                  0                     0               0
#>   AW2               0                  0                     1               0
#>   AW3               0                  0                     0               0
#>   AW4               0                  0                     1               0
#>   AW5               0                  0                     1               0

# Reproduce the summary table (Appendix S2) in the supplementary
# material of Neo et al. (2017)
aggregate(. ~ patch, data = sf::st_drop_geometry(novelforest_data$plots)[-1],
    FUN = mean, na.rm = TRUE) |>
  merge(novelforest_data$patches)
#>    patch  canopy litter nitrogen phosphorous potassium       dist type
#> 1     AD 91.6904  1.780    2.380      10.358     65.36  3.2807506  ABA
#> 2     AW 76.0540  2.908    2.870       8.798     66.62  3.7396397  WAS
#> 3    BBE 51.6364  5.960    1.932       7.594     64.12  1.2709380  WAS
#> 4    BBH 86.9360  5.480    1.162       4.872     38.92  2.9673487  ABA
#> 5    BBN 87.1456  3.480    1.540       9.498     49.08  1.0196463  ABA
#> 6    BBT 86.0848  3.200    1.988       4.250     46.30  1.7308310  ABA
#> 7     BR 46.0012  7.120    2.282      10.736    101.08 11.1164994  WAS
#> 8     CL 94.5504  4.500    2.408      91.812     75.34  2.2207678  ABA
#> 9     DO 79.1168  4.040    1.750      85.410     94.08  3.6578393  ABA
#> 10    GA 83.3808  3.700    2.324       9.584     33.26  0.2251064  ABA
#> 11    IK 70.3164  5.132    2.100       9.194     99.56 11.3692258  WAS
#> 12    JR 54.8760  4.180    1.960       4.758     67.26  3.4125817  WAS
#> 13    KJ 37.9172  4.060    2.828       6.000     68.40  3.7959059  WAS
#> 14    LD 79.0648  2.040    3.668      14.076     70.68  1.7867715  ABA
#> 15    PO 87.7280  2.880    2.058      25.178     64.92 10.4806218  ABA
#> 16    PT 84.1164  2.240    2.912      28.740    104.06  0.4624804  WAS
#> 17    SA 86.2884  2.784    2.408       9.840     62.18  2.5300102  WAS
#> 18    SJ 73.9604  5.068    1.736       7.072     56.64  1.8532589  ABA
#> 19    TP 60.8760  4.568    1.946       7.368     75.16 10.4218470  WAS
#> 20    UT 85.7832  2.380    2.492      16.820     45.96  0.4374091  ABA
#>         size
#> 1  38.153295
#> 2  21.449640
#> 3   9.661922
#> 4  32.870323
#> 5  28.094473
#> 6  34.155915
#> 7  16.285152
#> 8  37.479339
#> 9  16.876638
#> 10 57.411090
#> 11 23.201688
#> 12 19.663564
#> 13 11.411731
#> 14 26.322790
#> 15 10.378767
#> 16 12.518408
#> 17  9.619246
#> 18  7.587687
#> 19 21.281870
#> 20 56.446318