Backtransform scaled predictors

Backtransform scaled predictors in the novelforestSG model to their original scales. This is done by first back-scaling to the log-scale, and then backtranformed to the original scale (the predictors were log-transformed as described in Lai et al. 2021).

backtransform(data = novelforest$data)

Arguments

data
Defaults to the data object in novelforestSG, but could also be another dataframe with the same predictor names, should you wish to (back)scale and (back)center using the same means and standard deviations for any reason.

Value

A backtransformed dataset with predictors at their original scales.

References

Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W. and Chong, K.Y. (2021) Decoupled responses of native and exotic tree diversities to distance from old-growth forest and soil phosphorous in novel secondary forests. Applied Vegetation Science. doi: 10.1111/avsc.12548

Examples

# argument could be left blank
backtransform()

#>    SD_N_0        dist      size nitrogen phosphorous potassium patch   SD_N_2
#> 1       2  3293.80000 38.153295     2730       17.40      59.1    AD 1.600000
#> 2       8  3174.00000 38.153295     1750        9.32      54.7    AD 5.128205
#> 3       2  3262.35305 38.153295     2590       12.40      82.5    AD 1.600000
#> 4       1  3315.60000 38.153295     2310        6.16      76.6    AD 1.000000
#> 5       0  3358.00000 38.153295     2520        6.51      53.9    AD 0.000000
#> 6       6  3671.50346 21.449640     2660       23.60      71.2    AW 4.121951
#> 7       5  3809.20000 21.449640     3360        7.32      89.3    AW 2.648649
#> 8       8  3605.01500 21.449640     2730        4.43      58.2    AW 4.232000
#> 9       3  3816.27174 21.449640     2590        4.99      56.0    AW 1.588235
#> 10      5  3796.20830 21.449640     3010        3.65      58.4    AW 2.000000
#> 11      6  1170.18079  9.661922     1400        8.87      34.6   BBE 4.571429
#> 12      7  1203.70564  9.661922     3780       12.40     139.0   BBE 4.444444
#> 13      9  1400.49574  9.661922     1470        9.68      68.6   BBE 7.117647
#> 14      8  1148.04050  9.661922     1400        2.40      46.4   BBE 1.838906
#> 15      4  1432.26729  9.661922     1610        4.62      32.0   BBE 3.266667
#> 16      7  2979.88286 32.870323      910        3.12      26.4   BBH 3.082192
#> 17      3  2983.43471 32.870323      840        3.48      26.3   BBH 3.000000
#> 18      6  2835.94971 32.870323     1190        6.64      47.1   BBH 4.840000
#> 19      7  3245.13892 32.870323     1190        3.12      38.8   BBH 2.232068
#> 20      2  2792.33715 32.870323     1680        8.00      56.0   BBH 2.000000
#> 21      6   972.48657 28.094473     1540        7.77      53.8   BBN 4.481481
#> 22      4  1089.66821 28.094473     1330       12.50      50.9   BBN 2.909091
#> 23      4  1136.96241 28.094473     1190       11.50      27.1   BBN 2.682540
#> 24      5   975.69924 28.094473     1470        9.26      65.4   BBN 2.864407
#> 25      5   923.41499 28.094473     2170        6.46      48.2   BBN 4.500000
#> 26      4  1612.35387 34.155915     1680        4.69      54.3   BBT 3.571429
#> 27      6  1727.98669 34.155915     1750        3.04      29.6   BBT 4.500000
#> 28      5  1679.31587 34.155915     2100        2.31      35.3   BBT 2.813953
#> 29      4  1744.61935 34.155915     2240        6.03      60.6   BBT 2.578947
#> 30      3  1889.87899 34.155915     2170        5.18      51.7   BBT 2.666667
#> 31      2 11131.93092 16.285152     2730       14.30      82.7    BR 2.000000
#> 32      1 11044.50207 16.285152     2660        6.79      82.7    BR 1.000000
#> 33      1 11085.31864 16.285152     2100       11.40     102.0    BR 1.000000
#> 34      4  2165.23060 37.479339     4690      131.60     119.6    CL 1.603960
#> 35      3  2326.24577 37.479339     1680        8.06      49.1    CL 1.851852
#> 36      3  2244.63418 37.479339     1960       87.40      91.2    CL 1.843658
#> 37      5  2211.07339 37.479339     1820      203.10      32.4    CL 3.125000
#> 38      6  2156.65498 37.479339     1890       28.90      84.4    CL 4.666667
#> 39      4  3674.12884 16.876638     1960      118.20      96.9    DO 2.909091
#> 40     10  3553.79875 16.876638     1260       12.10      98.3    DO 5.236364
#> 41      5  3528.15113 16.876638     1820        8.75      94.1    DO 3.650667
#> 42      3  3800.82368 16.876638     1680       92.10      87.6    DO 2.666667
#> 43      4  3732.29395 16.876638     2030      195.90      93.5    DO 2.909091
#> 44      8   161.59270 57.411090     2030        9.44      20.8    GA 6.250000
#> 45      6   486.30250 57.411090     2100       10.90      20.4    GA 2.964912
#> 46     12    83.03879 57.411090     1820        8.11      34.3    GA 4.370629
#> 47     10   220.61732 57.411090     2240        8.87      40.4    GA 8.166667
#> 48     11   173.98046 57.411090     3430       10.60      50.4    GA 7.810811
#> 49      0 11324.27196 23.201688      840        3.11      56.0    IK 0.000000
#> 50      2 11550.16947 23.201688     2590       15.30     124.0    IK 1.600000
#> 51      1 11337.00505 23.201688     2870       10.20     104.0    IK 1.000000
#> 52      0 11224.35878 23.201688     3010       13.40      79.8    IK 0.000000
#> 53      2 11410.32390 23.201688     1190        3.96     134.0    IK 2.000000
#> 54      1  3373.95293 19.663564     2940        1.56      67.4    JR 1.000000
#> 55      7  3532.01519 19.663564     1610        2.75      61.0    JR 3.983607
#> 56      3  3392.82015 19.663564     2380        5.48      78.9    JR 2.272727
#> 57      3  3481.92487 19.663564     1260        8.16      58.2    JR 1.741935
#> 58      5  3282.19517 19.663564     1610        5.84      70.8    JR 3.225806
#> 59      4  3663.77731 11.411731     2800        9.10      67.7    KJ 4.000000
#> 60      3  3686.59616 11.411731     4410        3.76      82.0    KJ 3.000000
#> 61      2  3791.15012 11.411731     2730        4.42      66.1    KJ 2.000000
#> 62      1  3901.02330 11.411731     1680        7.52      42.5    KJ 1.000000
#> 63      0  3936.98281 11.411731     2520        5.20      83.7    KJ 0.000000
#> 64      5  1851.21721 26.322790     2520       24.60      61.7    LD 2.469388
#> 65      8  1771.65375 26.322790     1610        6.66      55.5    LD 6.736842
#> 66      9  1769.11762 26.322790     1960        5.22      60.1    LD 6.428571
#> 67      3  1778.07131 26.322790     5110       13.40      72.1    LD 2.666667
#> 68      5  1763.79783 26.322790     7140       20.50     104.0    LD 5.000000
#> 69      6 10491.70557 10.378767     1120       24.20      23.8    PO 4.787234
#> 70      7 10504.59639 10.378767     2380       38.90      46.7    PO 2.084599
#> 71      2 10464.74311 10.378767     3010       20.20     142.0    PO 1.882353
#> 72      1 10430.95950 10.378767     2170        8.09      53.9    PO 1.000000
#> 73      1 10511.10423 10.378767     1610       34.50      58.2    PO 1.000000
#> 74      3   387.57544 12.518408     3080       71.70     101.0    PT 2.739726
#> 75      8   516.80004 12.518408     2590       14.00      72.7    PT 3.881988
#> 76      0   613.13049 12.518408     2940       14.70     104.0    PT 0.000000
#> 77      4   417.30804 12.518408     2450       13.50     147.0    PT 3.846154
#> 78      2   377.58782 12.518408     3500       29.80      95.6    PT 1.724138
#> 79      5  2413.39903  9.619246     2800       12.10      74.9    SA 3.446809
#> 80      7  2476.27139  9.619246     2100        8.01      44.0    SA 5.555556
#> 81      5  2635.67653  9.619246     2660       18.10      62.0    SA 3.595745
#> 82      8  2543.39869  9.619246     2100        3.92      64.2    SA 3.531073
#> 83      4  2581.30537  9.619246     2380        7.07      65.8    SA 2.941176
#> 84      5  1681.38648  7.587687     1330        9.37      45.5    SJ 2.000000
#> 85      4  1991.68586  7.587687     1400        6.55      50.4    SJ 3.521739
#> 86      2  1883.03488  7.587687      630        2.89      37.5    SJ 1.342282
#> 87      0  1752.26492  7.587687     3290        8.55      90.9    SJ 0.000000
#> 88      4  1957.92227  7.587687     2030        8.00      58.9    SJ 3.130435
#> 89      2 10649.04044 21.281870     2240        5.25      99.9    TP 1.600000
#> 90      2 10306.68189 21.281870     2450       13.70     113.0    TP 1.800000
#> 91      4 10336.70931 21.281870     2380        5.72      58.1    TP 2.076923
#> 92      2 10587.60737 21.281870     1820        4.14      30.2    TP 2.000000
#> 93      5   396.59359 56.446318     2100       15.10      35.6    UT 2.500000
#> 94      7   520.31021 56.446318     1750        8.60      28.9    UT 6.000000
#> 95      9   554.11641 56.446318     2170       30.70      46.4    UT 4.389610
#> 96      5   226.64739 56.446318     3150       17.30      48.2    UT 3.521739
#> 97      6   489.37793 56.446318     3290       12.40      70.7    UT 5.142857
#>    SD_E_0   SD_E_2     FD_N_0     FD_N_2    FD_E_0    FD_E_2
#> 1       3 1.549356   6.068161   6.068161 19.061061 14.349844
#> 2       2 1.124514 176.741963  80.528861  6.345825  6.345825
#> 3       2 1.045431   6.276896   6.276896  6.351461  6.351461
#> 4       1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 5       3 2.461538   0.000000   0.000000 19.061061 13.579309
#> 6       2 1.800000  95.149105  54.255080  6.345825  6.345825
#> 7       3 2.133333  62.773106  34.887145 19.167086 14.356864
#> 8       2 1.800000 177.033677  56.810206  6.345825  6.345825
#> 9       2 2.000000  19.299513  14.826368  6.396550  6.396550
#> 10      2 1.045431  62.215777  29.702331  6.396550  6.396550
#> 11      3 1.119171  94.236570  71.343714 19.356083 13.158064
#> 12      1 1.000000 135.157197  62.094609  0.000000  0.000000
#> 13      2 1.198020 231.052893 173.608806  6.391431  6.391431
#> 14      0 0.000000 176.327692  41.129042  0.000000  0.000000
#> 15      1 1.000000  38.250256  29.217622  0.000000  0.000000
#> 16      1 1.000000 134.058465  60.809027  0.000000  0.000000
#> 17      1 1.000000  18.793740  18.793740  0.000000  0.000000
#> 18      1 1.000000  94.236018  65.252739  0.000000  0.000000
#> 19      1 1.000000 132.295135  48.953894  0.000000  0.000000
#> 20      2 1.882353   6.274281   6.274281  6.387492  6.387492
#> 21      6 5.333333  96.881205  65.444844 95.614574 79.645150
#> 22      2 1.132743  38.786866  27.265040  6.202454  6.202454
#> 23      6 3.321839  38.858853  17.927659 95.705907 40.645428
#> 24      4 1.228879  64.043986  37.015475 37.592069 18.014955
#> 25      1 1.000000  63.029816  56.151139  0.000000  0.000000
#> 26      1 1.000000  38.367889  34.545024  0.000000  0.000000
#> 27      1 1.000000  96.190066  60.103087  0.000000  0.000000
#> 28      1 1.000000  62.998358  35.832949  0.000000  0.000000
#> 29      1 1.000000  38.378461  28.285722  0.000000  0.000000
#> 30      1 1.000000  19.063042  17.670762  0.000000  0.000000
#> 31      2 1.470588   6.120796   6.120796  6.201227  6.201227
#> 32      4 1.640777   0.000000   0.000000 38.109678 22.855183
#> 33      0 0.000000   0.000000   0.000000  0.000000  0.000000
#> 34      3 1.945946  37.438460  19.611355 19.356083 12.113701
#> 35      1 1.000000  18.711982  13.191554  0.000000  0.000000
#> 36      3 1.457831  18.429523   8.680781 19.206442 14.038891
#> 37      2 1.180328  64.619019  39.382151  6.202454  6.202454
#> 38      3 1.851852  95.373958  61.895365 19.338876 13.616616
#> 39      2 1.219512  37.445985  26.193978  6.371743  6.371743
#> 40      0 0.000000 287.339299 128.648669  0.000000  0.000000
#> 41      1 1.000000  64.017472  35.527770  0.000000  0.000000
#> 42      4 1.923077  19.242167  17.790449 38.168703 24.714305
#> 43      2 1.923077  38.094578  26.578393  6.371743  6.371743
#> 44      1 1.000000 176.981404 132.736059  0.000000  0.000000
#> 45      1 1.000000  94.720546  47.585811  0.000000  0.000000
#> 46      1 1.000000 420.347047 129.604871  0.000000  0.000000
#> 47      1 1.000000 283.387544 204.311674  0.000000  0.000000
#> 48      1 1.000000 346.268906 198.139140  0.000000  0.000000
#> 49      3 1.135255   0.000000   0.000000 19.085394 13.202412
#> 50      2 1.800000   6.301612   6.301612  6.396550  6.396550
#> 51      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 52      3 2.000000   0.000000   0.000000 19.041104 15.553319
#> 53      1 1.000000   6.309702   6.309702  0.000000  0.000000
#> 54      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 55      1 1.000000 132.228700  49.129541  0.000000  0.000000
#> 56      3 2.086420  19.291174  16.513704 18.806499 11.072302
#> 57      1 1.000000  19.094310  10.467885  0.000000  0.000000
#> 58      1 1.000000  63.622635  27.656591  0.000000  0.000000
#> 59      3 3.000000  38.376026  38.376026 19.071534 19.071534
#> 60      4 4.000000  19.268292  19.268292 37.983308 37.983308
#> 61      2 2.000000   6.526282   6.526282  6.327744  6.327744
#> 62      4 4.000000   0.000000   0.000000 37.946456 37.946456
#> 63      3 1.975610   0.000000   0.000000 19.071534 13.834141
#> 64      6 5.260870  62.748522  27.612225 94.613170 76.003687
#> 65      2 1.396807 176.829594 128.857883  6.371743  6.371743
#> 66      2 1.198020 226.360116 130.991195  6.387492  6.387492
#> 67      3 2.102804  19.062079  17.639762 19.269549 10.455300
#> 68      3 1.515152  63.783967  63.783967 19.004983 14.242388
#> 69      2 1.172608  94.784655  62.590512  6.202454  6.202454
#> 70      0 0.000000 130.886011  43.247232  0.000000  0.000000
#> 71      3 1.800000   6.276896   6.276896 18.937631 13.235973
#> 72      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 73      3 1.546828   0.000000   0.000000 18.937631  8.841082
#> 74      1 1.000000  18.763870  16.670540  0.000000  0.000000
#> 75      2 1.445902 174.232562  49.457428  6.327744  6.327744
#> 76      2 1.724138   0.000000   0.000000  6.327744  6.327744
#> 77      2 1.280000  38.657762  36.614554  6.327744  6.327744
#> 78      2 1.800000   6.291827   6.291827  6.327744  6.327744
#> 79      2 1.710059  64.393372  33.364788  6.327744  6.327744
#> 80      2 1.882353 135.724596  91.230646  6.327744  6.327744
#> 81      2 1.800000  63.678893  36.142133  6.327744  6.327744
#> 82      1 1.000000 177.309397  62.124638  0.000000  0.000000
#> 83      1 1.000000  37.763198  21.365859  0.000000  0.000000
#> 84      2 1.689655  63.518026  30.329428  6.387492  6.387492
#> 85      4 2.769231  38.179587  30.619519 37.592069 18.300971
#> 86      2 1.470588   6.262615   6.262615  6.396550  6.396550
#> 87      2 2.000000   0.000000   0.000000  6.483252  6.483252
#> 88      1 1.000000  38.604423  25.415840  0.000000  0.000000
#> 89      1 1.000000   6.331174   6.331174  0.000000  0.000000
#> 90      2 1.324324   6.474740   6.474740  6.201227  6.201227
#> 91      3 1.684211  38.193122  25.473886 18.534312 14.389511
#> 92      2 1.753425   6.344339   6.344339  6.201227  6.201227
#> 93      1 1.000000  62.774729  37.542058  0.000000  0.000000
#> 94      1 1.000000 133.193222 102.797463  0.000000  0.000000
#> 95      3 3.000000 226.585433  69.508383 18.504130 18.504130
#> 96      2 1.600000  63.086499  41.744229  6.428676  6.428676
#> 97      5 3.595745  95.965939  73.268482 63.746937 36.094626