Novel Forest Dataset

The dataset used in Lai et al. (2021), who examined the diversity--environment relationships of native and exotic species in two types of post-cultivation novel secondary forests in Singapore.

novelforest_data

Format

A data.frame containing the following response variables:

SD_N_0

first-order native taxonomic diversity, i.e., species richness

SD_N_2

second-order native taxonomic diversity, i.e., inverse Simpson index

SD_E_0

first-order exotic taxonomic diversity

SD_E_2

second-order exotic taxonomic diversity

FD_N_0

first-order native functional diversity

FD_N_2

second-order native functional diversity

FD_E_0

first-order exotic functional diversity

FD_E_2

second-order exotic functional diversity,

and the following explanatory variables (and measurement units if you backtransform them using backtransform):

dist

Distance to old-growth forests (m)

size

Patch area (km^2)

nitrogen

Total soil nitrogen (mg/kg)

phosphorous

Total extractable soil phosphorous (mg/kg)

potassium

Total extractable soil potassium (mg/kg)

patch

Forest patch ID

Note that all explanatory variables were log-transformed and standardised to zero mean and unit standard deviations. Use backtransform to obtain the variables in their original scales. See Lai et al. (2021) for more details on model building and data collection.

References

Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W. and Chong, K.Y. (2021) Decoupled responses of native and exotic tree diversities to distance from old-growth forest and soil phosphorous in novel secondary forests. Applied Vegetation Science, 24, e12548. https://doi.org/10.1111/avsc.12548

See also

Examples

# To access the dataset
novelforest_data

#>    SD_N_0        dist        size    nitrogen phosphorous   potassium patch
#> 1       2  0.34018403  1.01559837  0.67459639  0.57571244 -0.05023288    AD
#> 2       8  0.30610447  1.01559837 -0.44642115 -0.08104609 -0.22461032    AD
#> 3       2  0.33135980  1.01559837  0.54188573  0.21932955  0.70159236    AD
#> 4       1  0.34625195  1.01559837  0.25346635 -0.51665575  0.53435074    AD
#> 5       0  0.35794036  1.01559837  0.47281510 -0.45852057 -0.25781745    AD
#> 6       6  0.44004163  0.04359803  0.60911425  0.89633096  0.36958155    AW
#> 7       5  0.47390839  0.04359803  1.19803868 -0.33515406  0.88010791    AW
#> 8       8  0.42323121  0.04359803  0.67459639 -0.86346927 -0.08482016    AW
#> 9       3  0.47561449  0.04359803  0.54188573 -0.73824554 -0.17167092    AW
#> 10      5  0.47076580  0.04359803  0.92073519 -1.06720860 -0.07708810    AW
#> 11      6 -0.61174710 -1.30243313 -1.00894826 -0.13310625 -1.25691603   BBE
#> 12      7 -0.58576466 -1.30243313  1.49496033  0.21932955  1.87739357   BBE
#> 13      9 -0.44648062 -1.30243313 -0.88595215 -0.04117684  0.28573608   BBE
#> 14      8 -0.62931766 -1.30243313 -1.00894826 -1.50825971 -0.59552058   BBE
#> 15      4 -0.42584629 -1.30243313 -0.65661954 -0.81929128 -1.43298517   BBE
#> 16      7  0.24805440  0.76404822 -2.09491768 -1.23225799 -1.86657088   BBH
#> 17      3  0.24915015  0.76404822 -2.29669897 -1.11738261 -1.87512457   BBH
#> 18      6  0.20251560  0.76404822 -1.41864554 -0.43772030 -0.56177181   BBH
#> 19      7  0.32649330  0.76404822 -1.41864554 -1.23225799 -0.99869523   BBH
#> 20      2  0.18825991  0.76404822 -0.54933014 -0.24170548 -0.17167092   BBH
#> 21      6 -0.78197085  0.49907107 -0.76867889 -0.27239316 -0.26200296   BBN
#> 22      4 -0.67731816  0.49907107 -1.13825457  0.22777923 -0.38689241   BBN
#> 23      4 -0.63823685  0.49907107 -1.41864554  0.14006352 -1.80758713   BBN
#> 24      5 -0.77893710  0.49907107 -0.88595215 -0.08784038  0.17806697   BBN
#> 25      5 -0.82959808  0.49907107  0.09585751 -0.46663147 -0.50973839   BBN
#> 26      4 -0.31690329  0.82880093 -0.54933014 -0.80347175 -0.24115273   BBT
#> 27      6 -0.25319324  0.82880093 -0.44642115 -1.25958366 -1.60870216   BBT
#> 28      5 -0.27947361  0.82880093  0.01319697 -1.54846763 -1.21177217   BBT
#> 29      4 -0.24438167  0.82880093  0.17589345 -0.53909422  0.00625863   BBT
#> 30      3 -0.17081588  0.82880093  0.09585751 -0.69893398 -0.35174326   BBT
#> 31      2  1.46034726 -0.42130821  0.67459639  0.36930292  0.70704973    BR
#> 32      1  1.45309440 -0.42130821  0.60911425 -0.41422016  0.70704973    BR
#> 33      1  1.45648755 -0.42130821  0.01319697  0.13087587  1.17981108    BR
#> 34      4 -0.04570421  0.98551824  2.03874446  2.70417842  1.53858587    CL
#> 35      3  0.02027511  0.98551824 -0.54933014 -0.23384507 -0.46804133    CL
#> 36      3 -0.01257541  0.98551824 -0.16072856  2.27363310  0.92756005    CL
#> 37      5 -0.02643236  0.98551824 -0.34754885  3.16066524 -1.40498614    CL
#> 38      6 -0.04935458  0.98551824 -0.25240849  1.10945654  0.75291150    CL
#> 39      4  0.44069915 -0.36109405 -0.16072856  2.59120752  1.06420147    DO
#> 40     10  0.41006932 -0.36109405 -1.27455373  0.19356551  1.09653246    DO
#> 41      5  0.40340678 -0.36109405 -0.34754885 -0.14743536  0.99811400    DO
#> 42      3  0.47188345 -0.36109405 -0.54933014  2.32873541  0.83678698    DO
#> 43      4  0.45514712 -0.36109405 -0.07226614  3.12269495  0.98369673    DO
#> 44      8 -2.43288531  1.70526856 -0.07226614 -0.06758775 -2.40392388    GA
#> 45      6 -1.41944585  1.70526856  0.01319697  0.08369401 -2.44769017    GA
#> 46     12 -3.04529065  1.70526856 -0.34754885 -0.22733931 -1.27654368    GA
#> 47     10 -2.14649141  1.70526856  0.17589345 -0.13310625 -0.90761644    GA
#> 48     11 -2.36494195  1.70526856  1.25001826  0.05433455 -0.40914228    GA
#> 49      0  1.47610486  0.17611782 -2.29669897 -1.23563513 -0.17167092    IK
#> 50      2  1.49427336  0.17611782  0.54188573  0.44040965  1.62001616    IK
#> 51      1  1.47713856  0.17611782  0.80066867  0.01386883  1.22357738    IK
#> 52      0  1.46795314  0.17611782  0.92073519  0.30091920  0.62659454    IK
#> 53      2  1.48306823  0.17611782 -1.41864554 -0.98145457  1.79482414    IK
#> 54      1  0.36229994 -0.10313850  0.86141668 -1.96143434  0.24596046    JR
#> 55      7  0.40441364 -0.10313850 -0.65661954 -1.36505151  0.02108693    JR
#> 56      3  0.36742939 -0.10313850  0.32872329 -0.63970751  0.60103026    JR
#> 57      3  0.39127523 -0.10313850 -1.27455373 -0.22087353 -0.08482016    JR
#> 58      5  0.33693748 -0.10313850 -0.65661954 -0.57277461  0.35688352    JR
#> 59      4  0.43810391 -1.02150327  0.73842056 -0.10617599  0.25597036    KJ
#> 60      3  0.44381514 -1.02150327  1.88356192 -1.03597346  0.68789084    KJ
#> 61      2  0.46953936 -1.02150327  0.67459639 -0.86584662  0.20206302    KJ
#> 62      1  0.49581880 -1.02150327 -0.54933014 -0.30679711 -0.79340200    KJ
#> 63      0  0.50425905 -1.02150327  0.47281510 -0.69488011  0.73414012    KJ
#> 64      5 -0.18982852  0.38913283  0.47281510  0.93998781  0.04680400    LD
#> 65      8 -0.23023720  0.38913283 -0.65661954 -0.43455645 -0.19188535    LD
#> 66      9 -0.23155490  0.38913283 -0.16072856 -0.69084179 -0.01241499    LD
#> 67      3 -0.22691122  0.38913283  2.25495579  0.30091920  0.39789321    LD
#> 68      5 -0.23432506  0.38913283  3.09823735  0.74818934  1.22357738    LD
#> 69      6  1.40586249 -1.18163961 -1.57147538  0.92274186 -2.10025099    PO
#> 70      7  1.40699198 -1.18163961  0.32872329  1.42205493 -0.58099490    PO
#> 71      2  1.40349556 -1.18163961  0.92073519  0.73268077  1.92552123    PO
#> 72      1  1.40052120 -1.18163961  0.09585751 -0.22993679 -0.25781745    PO
#> 73      1  1.40756167 -1.18163961 -0.65661954  1.29578069 -0.08482016    PO
#> 74      3 -1.62817709 -0.86528500  0.97868993  2.06533647  1.15760506    PT
#> 75      8 -1.36349621 -0.86528500  0.54188573  0.34699864  0.41657199    PT
#> 76      0 -1.20627492 -0.86528500  0.86141668  0.39832487  1.22357738    PT
#> 77      4 -1.56018763 -0.86528500  0.40179856  0.30874064  2.00351840    PT
#> 78      2 -1.65219175 -0.86528500  1.30094768  1.14171737  1.03375881    PT
#> 79      5  0.05410735 -1.30990447  0.73842056  0.19356551  0.48376620    SA
#> 80      7  0.07776372 -1.30990447  0.01319697 -0.24039133 -0.71522445    SA
#> 81      5  0.13514903 -1.30990447  0.60911425  0.61720432  0.05773641    SA
#> 82      8  0.10236705 -1.30990447  0.01319697 -0.99213466  0.13632702    SA
#> 83      4  0.11597520 -1.30990447  0.32872329 -0.37171017  0.19181027    SA
#> 84      5 -0.27834014 -1.71031213 -1.13825457 -0.07541750 -0.63966788    SJ
#> 85      4 -0.12255293 -1.71031213 -1.00894826 -0.45207657 -0.40914228    SJ
#> 86      2 -0.17415310 -1.71031213 -3.02192255 -1.31281487 -1.07550636    SJ
#> 87      0 -0.24035938 -1.71031213  1.14496472 -0.17175966  0.92013370    SJ
#> 88      4 -0.13828006 -1.71031213 -0.07226614 -0.24170548 -0.05787320    SJ
#> 89      2  1.41955417  0.03034498  0.17589345 -0.68481324  1.13292305    TP
#> 90      2  1.38949609  0.03034498  0.40179856  0.32421120  1.41064357    TP
#> 91      4  1.39217206  0.03034498  0.32872329 -0.59461579 -0.08869616    TP
#> 92      2  1.41423233  0.03034498 -0.34754885 -0.93469225 -1.56347202    TP
#> 93      5 -1.60701925  1.67666501  0.01319697  0.42656762 -1.19269821    UT
#> 94      7 -1.35726962  1.67666501 -0.44642115 -0.16562565 -1.66264406    UT
#> 95      9 -1.29936584  1.67666501  0.09585751  1.17301824 -0.59552058    UT
#> 96      5 -2.12168702  1.67666501  1.03534221  0.56964915 -0.50973839    UT
#> 97      6 -1.41364697  1.67666501  1.14496472  0.21932955  0.35369781    UT
#>      SD_N_2 SD_E_0   SD_E_2     FD_N_0     FD_N_2    FD_E_0    FD_E_2
#> 1  1.600000      3 1.549356   6.068161   6.068161 19.061061 14.349844
#> 2  5.128205      2 1.124514 176.741963  80.528861  6.345825  6.345825
#> 3  1.600000      2 1.045431   6.276896   6.276896  6.351461  6.351461
#> 4  1.000000      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 5  0.000000      3 2.461538   0.000000   0.000000 19.061061 13.579309
#> 6  4.121951      2 1.800000  95.149105  54.255080  6.345825  6.345825
#> 7  2.648649      3 2.133333  62.773106  34.887145 19.167086 14.356864
#> 8  4.232000      2 1.800000 177.033677  56.810206  6.345825  6.345825
#> 9  1.588235      2 2.000000  19.299513  14.826368  6.396550  6.396550
#> 10 2.000000      2 1.045431  62.215777  29.702331  6.396550  6.396550
#> 11 4.571429      3 1.119171  94.236570  71.343714 19.356083 13.158064
#> 12 4.444444      1 1.000000 135.157197  62.094609  0.000000  0.000000
#> 13 7.117647      2 1.198020 231.052893 173.608806  6.391431  6.391431
#> 14 1.838906      0 0.000000 176.327692  41.129042  0.000000  0.000000
#> 15 3.266667      1 1.000000  38.250256  29.217622  0.000000  0.000000
#> 16 3.082192      1 1.000000 134.058465  60.809027  0.000000  0.000000
#> 17 3.000000      1 1.000000  18.793740  18.793740  0.000000  0.000000
#> 18 4.840000      1 1.000000  94.236018  65.252739  0.000000  0.000000
#> 19 2.232068      1 1.000000 132.295135  48.953894  0.000000  0.000000
#> 20 2.000000      2 1.882353   6.274281   6.274281  6.387492  6.387492
#> 21 4.481481      6 5.333333  96.881205  65.444844 95.614574 79.645150
#> 22 2.909091      2 1.132743  38.786866  27.265040  6.202454  6.202454
#> 23 2.682540      6 3.321839  38.858853  17.927659 95.705907 40.645428
#> 24 2.864407      4 1.228879  64.043986  37.015475 37.592069 18.014955
#> 25 4.500000      1 1.000000  63.029816  56.151139  0.000000  0.000000
#> 26 3.571429      1 1.000000  38.367889  34.545024  0.000000  0.000000
#> 27 4.500000      1 1.000000  96.190066  60.103087  0.000000  0.000000
#> 28 2.813953      1 1.000000  62.998358  35.832949  0.000000  0.000000
#> 29 2.578947      1 1.000000  38.378461  28.285722  0.000000  0.000000
#> 30 2.666667      1 1.000000  19.063042  17.670762  0.000000  0.000000
#> 31 2.000000      2 1.470588   6.120796   6.120796  6.201227  6.201227
#> 32 1.000000      4 1.640777   0.000000   0.000000 38.109678 22.855183
#> 33 1.000000      0 0.000000   0.000000   0.000000  0.000000  0.000000
#> 34 1.603960      3 1.945946  37.438460  19.611355 19.356083 12.113701
#> 35 1.851852      1 1.000000  18.711982  13.191554  0.000000  0.000000
#> 36 1.843658      3 1.457831  18.429523   8.680781 19.206442 14.038891
#> 37 3.125000      2 1.180328  64.619019  39.382151  6.202454  6.202454
#> 38 4.666667      3 1.851852  95.373958  61.895365 19.338876 13.616616
#> 39 2.909091      2 1.219512  37.445985  26.193978  6.371743  6.371743
#> 40 5.236364      0 0.000000 287.339299 128.648669  0.000000  0.000000
#> 41 3.650667      1 1.000000  64.017472  35.527770  0.000000  0.000000
#> 42 2.666667      4 1.923077  19.242167  17.790449 38.168703 24.714305
#> 43 2.909091      2 1.923077  38.094578  26.578393  6.371743  6.371743
#> 44 6.250000      1 1.000000 176.981404 132.736059  0.000000  0.000000
#> 45 2.964912      1 1.000000  94.720546  47.585811  0.000000  0.000000
#> 46 4.370629      1 1.000000 420.347047 129.604871  0.000000  0.000000
#> 47 8.166667      1 1.000000 283.387544 204.311674  0.000000  0.000000
#> 48 7.810811      1 1.000000 346.268906 198.139140  0.000000  0.000000
#> 49 0.000000      3 1.135255   0.000000   0.000000 19.085394 13.202412
#> 50 1.600000      2 1.800000   6.301612   6.301612  6.396550  6.396550
#> 51 1.000000      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 52 0.000000      3 2.000000   0.000000   0.000000 19.041104 15.553319
#> 53 2.000000      1 1.000000   6.309702   6.309702  0.000000  0.000000
#> 54 1.000000      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 55 3.983607      1 1.000000 132.228700  49.129541  0.000000  0.000000
#> 56 2.272727      3 2.086420  19.291174  16.513704 18.806499 11.072302
#> 57 1.741935      1 1.000000  19.094310  10.467885  0.000000  0.000000
#> 58 3.225806      1 1.000000  63.622635  27.656591  0.000000  0.000000
#> 59 4.000000      3 3.000000  38.376026  38.376026 19.071534 19.071534
#> 60 3.000000      4 4.000000  19.268292  19.268292 37.983308 37.983308
#> 61 2.000000      2 2.000000   6.526282   6.526282  6.327744  6.327744
#> 62 1.000000      4 4.000000   0.000000   0.000000 37.946456 37.946456
#> 63 0.000000      3 1.975610   0.000000   0.000000 19.071534 13.834141
#> 64 2.469388      6 5.260870  62.748522  27.612225 94.613170 76.003687
#> 65 6.736842      2 1.396807 176.829594 128.857883  6.371743  6.371743
#> 66 6.428571      2 1.198020 226.360116 130.991195  6.387492  6.387492
#> 67 2.666667      3 2.102804  19.062079  17.639762 19.269549 10.455300
#> 68 5.000000      3 1.515152  63.783967  63.783967 19.004983 14.242388
#> 69 4.787234      2 1.172608  94.784655  62.590512  6.202454  6.202454
#> 70 2.084599      0 0.000000 130.886011  43.247232  0.000000  0.000000
#> 71 1.882353      3 1.800000   6.276896   6.276896 18.937631 13.235973
#> 72 1.000000      1 1.000000   0.000000   0.000000  0.000000  0.000000
#> 73 1.000000      3 1.546828   0.000000   0.000000 18.937631  8.841082
#> 74 2.739726      1 1.000000  18.763870  16.670540  0.000000  0.000000
#> 75 3.881988      2 1.445902 174.232562  49.457428  6.327744  6.327744
#> 76 0.000000      2 1.724138   0.000000   0.000000  6.327744  6.327744
#> 77 3.846154      2 1.280000  38.657762  36.614554  6.327744  6.327744
#> 78 1.724138      2 1.800000   6.291827   6.291827  6.327744  6.327744
#> 79 3.446809      2 1.710059  64.393372  33.364788  6.327744  6.327744
#> 80 5.555556      2 1.882353 135.724596  91.230646  6.327744  6.327744
#> 81 3.595745      2 1.800000  63.678893  36.142133  6.327744  6.327744
#> 82 3.531073      1 1.000000 177.309397  62.124638  0.000000  0.000000
#> 83 2.941176      1 1.000000  37.763198  21.365859  0.000000  0.000000
#> 84 2.000000      2 1.689655  63.518026  30.329428  6.387492  6.387492
#> 85 3.521739      4 2.769231  38.179587  30.619519 37.592069 18.300971
#> 86 1.342282      2 1.470588   6.262615   6.262615  6.396550  6.396550
#> 87 0.000000      2 2.000000   0.000000   0.000000  6.483252  6.483252
#> 88 3.130435      1 1.000000  38.604423  25.415840  0.000000  0.000000
#> 89 1.600000      1 1.000000   6.331174   6.331174  0.000000  0.000000
#> 90 1.800000      2 1.324324   6.474740   6.474740  6.201227  6.201227
#> 91 2.076923      3 1.684211  38.193122  25.473886 18.534312 14.389511
#> 92 2.000000      2 1.753425   6.344339   6.344339  6.201227  6.201227
#> 93 2.500000      1 1.000000  62.774729  37.542058  0.000000  0.000000
#> 94 6.000000      1 1.000000 133.193222 102.797463  0.000000  0.000000
#> 95 4.389610      3 3.000000 226.585433  69.508383 18.504130 18.504130
#> 96 3.521739      2 1.600000  63.086499  41.744229  6.428676  6.428676
#> 97 5.142857      5 3.595745  95.965939  73.268482 63.746937 36.094626

summary(novelforest_data)

#>      SD_N_0            dist              size            nitrogen      
#>  Min.   : 0.000   Min.   :-3.0453   Min.   :-1.7103   Min.   :-3.0219  
#>  1st Qu.: 2.000   1st Qu.:-0.4258   1st Qu.:-1.0215   1st Qu.:-0.5493  
#>  Median : 4.000   Median : 0.1160   Median : 0.0436   Median : 0.0132  
#>  Mean   : 4.402   Mean   : 0.0000   Mean   : 0.0000   Mean   : 0.0000  
#>  3rd Qu.: 6.000   3rd Qu.: 0.4551   3rd Qu.: 0.8288   3rd Qu.: 0.6091  
#>  Max.   :12.000   Max.   : 1.4943   Max.   : 1.7053   Max.   : 3.0982  
#>                                                                        
#>   phosphorous        potassium            patch        SD_N_2     
#>  Min.   :-1.9614   Min.   :-2.44769   AD     : 5   Min.   :0.000  
#>  1st Qu.:-0.6397   1st Qu.:-0.56177   AW     : 5   1st Qu.:1.839  
#>  Median :-0.1474   Median :-0.01242   BBE    : 5   Median :2.814  
#>  Mean   : 0.0000   Mean   : 0.00000   BBH    : 5   Mean   :2.980  
#>  3rd Qu.: 0.3470   3rd Qu.: 0.70705   BBN    : 5   3rd Qu.:4.000  
#>  Max.   : 3.1607   Max.   : 2.00352   BBT    : 5   Max.   :8.167  
#>                                       (Other):67                  
#>      SD_E_0          SD_E_2          FD_N_0            FD_N_2       
#>  Min.   :0.000   Min.   :0.000   Min.   :  0.000   Min.   :  0.000  
#>  1st Qu.:1.000   1st Qu.:1.000   1st Qu.:  6.344   1st Qu.:  6.344  
#>  Median :2.000   Median :1.220   Median : 38.604   Median : 28.286  
#>  Mean   :2.031   Mean   :1.539   Mean   : 70.365   Mean   : 39.320  
#>  3rd Qu.:3.000   3rd Qu.:1.800   3rd Qu.: 95.374   3rd Qu.: 56.151  
#>  Max.   :6.000   Max.   :5.333   Max.   :420.347   Max.   :204.312  
#>                                                                     
#>      FD_E_0           FD_E_2      
#>  Min.   : 0.000   Min.   : 0.000  
#>  1st Qu.: 0.000   1st Qu.: 0.000  
#>  Median : 6.328   Median : 6.328  
#>  Mean   :11.508   Mean   : 8.621  
#>  3rd Qu.:18.938   3rd Qu.:12.114  
#>  Max.   :95.706   Max.   :79.645  
#>