Novel Forest Dataset

The dataset used in Lai et al. (2021), who examined the diversity--environment relationships of native and exotic species in two types of post-cultivation novel secondary forests in Singapore.

novelforest_data

Format

A data.frame containing the following response variables:

SD_N_0

first-order native taxonomic diversity, i.e., species richness

SD_N_2

second-order native taxonomic diversity, i.e., inverse Simpson index

SD_E_0

first-order exotic taxonomic diversity

SD_E_2

second-order exotic taxonomic diversity

FD_N_0

first-order native functional diversity

FD_N_2

second-order native functional diversity

FD_E_0

first-order exotic functional diversity

FD_E_2

second-order exotic functional diversity,

and the following explanatory variables (and measurement units if you backtransform them using backtransform):

dist

Distance to old-growth forests (m)

size

Patch area (km^2)

nitrogen

Total soil nitrogen (mg/kg)

phosphorous

Total extractable soil phosphorous (mg/kg)

potassium

Total extractable soil potassium (mg/kg)

patch

Forest patch ID

Note that all explanatory variables were log-transformed and standardised to zero mean and unit standard deviations. Use backtransform to obtain the variables in their original scales. See Lai et al. (2021) for more details on model building and data collection.

References

Lai, H.R., Tan, G.S.Y., Neo, L., Kee, C.Y., Yee, A.T.K., Tan, H.T.W. and Chong, K.Y. (2021) Decoupled responses of native and exotic tree diversities to distance from old-growth forest and soil phosphorous in novel secondary forests. Applied Vegetation Science, 24, e12548. doi: 10.1111/avsc.12548

See also

Examples

# To access the dataset
head(novelforest_data)

#>   SD_N_0      dist       size   nitrogen phosphorous   potassium patch   SD_N_2
#> 1      2 0.3401840 1.01559837  0.6745964  0.57571244 -0.05023288    AD 1.600000
#> 2      8 0.3061045 1.01559837 -0.4464211 -0.08104609 -0.22461032    AD 5.128205
#> 3      2 0.3313598 1.01559837  0.5418857  0.21932955  0.70159236    AD 1.600000
#> 4      1 0.3462519 1.01559837  0.2534663 -0.51665575  0.53435074    AD 1.000000
#> 5      0 0.3579404 1.01559837  0.4728151 -0.45852057 -0.25781745    AD 0.000000
#> 6      6 0.4400416 0.04359803  0.6091143  0.89633096  0.36958155    AW 4.121951
#>   SD_E_0   SD_E_2     FD_N_0    FD_N_2    FD_E_0    FD_E_2
#> 1      3 1.549356   6.068161  6.068161 19.061061 14.349844
#> 2      2 1.124514 176.741963 80.528861  6.345825  6.345825
#> 3      2 1.045431   6.276896  6.276896  6.351461  6.351461
#> 4      1 1.000000   0.000000  0.000000  0.000000  0.000000
#> 5      3 2.461538   0.000000  0.000000 19.061061 13.579309
#> 6      2 1.800000  95.149105 54.255080  6.345825  6.345825